Populate fails if key doesn't exist
I noticed one bug in v2.2.0, if you try to populate across a collection of documents on a key that may or may not exist Mongorito throws an exception. The variable idsArray ends up being undefined and then any calls on it fail. 

I figured a simple fix would be to fallback to an empty array but let me know if you think a better solution fits.